+++ branches/trunk_2.0.0/smapi/src/gnmsgid.c
@@ -180,7 +180,7 @@ int main(int argc, char *argv[])
             break;
         }
         s = argv[i];
-        for(j=strlen(argv[i]); j>0; j--, s++)
+        for(j=(int)strlen(argv[i]); j>0; j--, s++)
             if (!isdigit((int)(*s)))
             {
                 fprintf(stderr, "Invalid <num> parameter ('%s')!\n", argv[i]);
