+++ branches/trunk_2.0.0/src/parsetpl.cpp
@@ -1570,7 +1570,7 @@ int Template::Save(cMSG & m)
     char * tmt;
 
     CHP = 77312;
-    m.SetBody(Body, strlen(Body));
+    m.SetBody(Body, (unsigned int)strlen(Body));
     CHP = 77314;
 
     while((tmt = strchr(m.Body(), '\n')) != NULL)
